Avoid removing test files during tests
Remove the pull pattern for the test file directory as there is no
reason to pull these files: the screenshots and connectivity
diagnostics are already covered by the other pull patterns.

Because the collector deletes the files on pull, this also causes test
failures where sometimes it pulls and deletes test files after a
test, but during the next test, due to collect-on-run-ended-only=false.
Also use true to avoid files getting pulled while tests are running.

This was originally added copying another test that used
FilePullerLogCollector, but is unsafe in this test.

Bug: 317602748
Test: atest
Flag: TEST_ONLY
Change-Id: Id54c61b3d1ec23dce7ba5768dc43dd910d79c44d