Add an opt-in mechanism for custom tabs in captive portals.
If the capport JSON file contains a field "x-android-use-custom-tabs"
with a value evaluating to true, the login app will use the custom
tabs even if the feature is not enabled yet.
Of course the feature still needs to be supported by the device –
it won't use it on a device without a browser with support.

Test: manual
Flag: EXEMPT opt-in feature
Change-Id: I5e56b4f2eb478c3af3f9624470fbd1a0053fe253