Give more time to UI to happen
On slow devices, instrumented UI can take a long time
to actually happen, animate and be detected. A longer
timeout does not affect the functionality of the tests
and will likely not substantially increase their run
time, but it will reduce the flake rate (if these
tests are indeed flaky because the device is too slow,
which is the most probable explanation here).

Bug: 424303208
Flag: EXEMPT test-only
Test: TH
Change-Id: Id063b4efdcae673cb7d9a75dea871757bd9359f3