Dismiss activity if private DNS is enabled on an active custom tab.
Addresses an edge case: when a CCT is open and the user navigates to
settings to set up private DNS, dismiss the CaptivePortalActivity as
the private DNS bypass logic only works on Android B and above. The
Webview will be triggered instead of the custom tab next time when the
portal is relaunched.

Bug: 368416967
Test: TH
Flag: EXEMPT protected by legacy flag `captive_portal_custom_tabs`
Change-Id: Idc5830232b79d0a8bf07d4773a4f579140b45877