Ensure CCT dismisses parent on `TAB_HIDDEN` event
When CCT receives a `TAB_HIDDEN` event and the parent activity is
resumed, both CCT and the app must be closed together. Because This
combination means CCT should be dismissed. Failing to do so leads to a
state where CCT closes, but the app remains open. This results in a poor
user experience, as navigating back displays an incomplete
`CaptivePortalLoginApp` screen with only the title visible.

Test: TH, manual
Flag: EXEMPT guarded by legacy flag
Change-Id: I0df5aba48e183c1e8faf93a799788e1a14bcc3b7