Properly either mock custom tabs, or use real ones
Custom tabs were half-mocked in tests, where the service connection and
callbacks were mocked, but the real intent was sent so the actual
browser was launched. Some tests depended on the mocking, and some
others on an actual browser to be launched.

Ensure mocking is either done fully, or not at all (so the real browser
is launched), depending on the need of tests.

Test: atest
Flag: EXEMPT test-only
Change-Id: I63de29a37e26f4f5aa29242c7b52efe5c1b2e892