Mock intents in test setUp
Instead of initializing intent mocking at activity creation time, and at
various times in different tests, do it in setUp().

This makes the setup easier to follow (test results do not depend on
when intent mocking was initialized), and it will allow using the same
setup for tests that use WebView, tests that use custom tabs, and tests
that use mocked custom tabs.

Test: atest
Flag: EXEMPT test-only
Change-Id: I3a0608f88a0454cbbf834180ad7e8a22f1b78674