Reland full custom tab in CaptivePortalLogin
Using a partial custom tab would be best as it allows showing additional
UI, in particular messages indicating that the user is currently logging
in to a portal. But this is unfortunately not doable due to current
technical limitations:

 - The TODO to recompute available space when the app changes size is
   infeasible with the current custom tab APIs, as there is no way to
   update the height of the tab. This prevents supporting screen
   rotation or any other screen resize event (the custom tab keeps the
   hardcoded height and overlaps with the status bar, or only uses a
   portion of the screen).
 - There are focus bugs with the partial tab where tapping on an already
   focused text field (for example to correct a typo) removes focus and
   dismisses the IME.

Not being able to rotate the screen would be a major regression compared
to the WebView implementation, so this reverts to a full custom tab
until rotation can be supported.

Bug: 433849360
Test: atest; opened portal manually and checked rotation, login flow,
  back button flow.
Flag: NONE legacy flag protected
Change-Id: I841b7a2b22163e1404f01646a6cd13d7d223b36b